requirements.txt
pytest 
responses
python-dotenv
pydantic
pydantic[email]